+++ test_suite/src/json_stream_traits_tests.cpp
@@ -16,7 +16,7 @@ using boost::numeric::ublas::matrix;
 namespace jsoncons
 {
     template <>
-    struct json_stream_traits<char,matrix<double>>
+    struct json_stream_traits<matrix<double>>
     {
         static void encode(const matrix<double>& val, json_output_handler& handler)
         {